Add files via upload
The main goal of this project is to dive deeper into some significant pandemics from human history, including Cholera, Dengue, and COVID-19 and rise up three questions to answer. Because we want to use data visualizations to visually present and explore the similarities and differences among these pandemics or diseases that happened on a large scale in different time periods.